refactor: Split functions to a new class to make them be testable (apply feedback)
